+++ nuxeo-drive-client/nxdrive/constants.py
@@ -1,0 +1,3 @@
+# coding: utf-8
+
+BUNDLE_IDENTIFIER = 'org.nuxeo.drive'