+++ nuxeo-drive-client/nxdrive/manager.py
@@ -379,7 +379,6 @@ class Manager(QtCore.QObject):
 
     def _handle_os(self):
         # Be sure to register os
-        self.osi.register_contextual_menu()
         self.osi.register_protocol_handlers()
         if self.get_auto_start():
             self.osi.register_startup()
@@ -873,9 +872,14 @@ class Manager(QtCore.QObject):
         return True
 
     def update_engine_path(self, uid, local_folder):
-        # Dont update the engine by itself, should be only used by engine.update_engine_path
+        # Dont update the engine by itself,
+        # should be only used by engine.update_engine_path
         if uid in self._engine_definitions:
+            # Unwatch old folder
+            self.osi.unwatch_folder(self._engine_definitions[uid].local_folder)
             self._engine_definitions[uid].local_folder = local_folder
+        # Watch new folder
+        self.osi.watch_folder(local_folder)
         self._dao.update_engine_path(uid, local_folder)
 
     def bind_engine(self, engine_type, local_folder, name, binder, starts=True):
@@ -909,6 +913,8 @@ class Manager(QtCore.QObject):
 
         uid = uuid.uuid1().hex
 
+        # Watch folder in the file explorer
+        self.osi.watch_folder(local_folder)
         # TODO Check that engine is not inside another or same position
         engine_def = self._dao.add_engine(engine_type, local_folder, uid, name)
         try:
@@ -940,6 +946,8 @@ class Manager(QtCore.QObject):
     def unbind_engine(self, uid):
         if self._engines is None:
             self.load()
+        # Unwatch folder
+        self.osi.unwatch_folder(self._engines[uid].local_folder)
         self._engines[uid].suspend()
         self._engines[uid].unbind()
         self._dao.delete_engine(uid)
@@ -993,10 +1001,6 @@ class Manager(QtCore.QObject):
                 return None
             return self.get_root_id(parent)
         return ref
-
-    @staticmethod
-    def get_cf_bundle_identifier():
-        return "org.nuxeo.drive"
 
     def open_metadata_window(self, file_path, application=None):
         """