+++ nuxeo-drive-client/nxdrive/utils.py
@@ -12,6 +12,8 @@ import time
 import urlparse
 from logging import getLogger
 
+import re
+
 from .options import Options
 
 DEVICE_DESCRIPTIONS = {
@@ -491,6 +493,66 @@ def simplify_url(url):
         new_parts[1] = parts.netloc[:-4]
 
     return urlparse.urlunsplit(new_parts).rstrip('/')
+
+
+def parse_protocol_url(url_string):
+    """
+    Parse URL for which Drive is registered as a protocol handler.
+
+    Return None if `url_string` is not a supported URL pattern or raise a
+    ValueError is the URL structure is invalid.
+    """
+    if not url_string.startswith('nxdrive://'):
+        return None
+
+    protocol_regex = [('nxdrive://(?P<cmd>edit)/(?P<scheme>\w*)/'
+                       '(?P<server>.*)/user/(?P<username>\w*)/repo/'
+                       '(?P<repo>\w*)/nxdocid/(?P<docid>(\d|[a-f]|-)*)/'
+                       'filename/(?P<filename>[^/]*)(/downloadUrl/'
+                       '(?P<download>.*)|)'),
+
+                      'nxdrive://(?P<cmd>share_link)/(?P<path>.*)',
+
+                      'nxdrive://(?P<cmd>access)/(?P<path>.*)',
+
+                      'nxdrive://(?P<cmd>trigger_watch)']
+
+    parsed_url = None
+    for regex in protocol_regex:
+        parsed_url = re.match(regex, url_string)
+        if parsed_url:
+            break
+
+    if not parsed_url:
+        raise ValueError(
+            'Unsupported command {!r} in protocol handler'.format(url_string))
+
+    parsed_url = parsed_url.groupdict()
+    cmd = parsed_url.get('cmd')
+    if cmd == 'edit':
+        return parse_edit_protocol(parsed_url, url_string)
+    if cmd in ('access', 'share_link'):
+        return dict(command=cmd, filepath=parsed_url.get('path'))
+    else:
+        return dict(command=cmd)
+
+
+def parse_edit_protocol(parsed_url, url_string):
+    """ Parse a `nxdrive://edit` URL for quick editing of Nuxeo documents. """
+    scheme = parsed_url.get('scheme')
+    if scheme not in ('http', 'https'):
+        raise ValueError(
+            'Invalid command {} : scheme should be http or https'.format(
+                url_string))
+
+    server_url = '{}://{}'.format(scheme, parsed_url.get('server'))
+
+    return dict(command='download_edit', server_url=server_url,
+                user=parsed_url.get('username'),
+                repo=parsed_url.get('repo'),
+                doc_id=parsed_url.get('docid'),
+                filename=parsed_url.get('filename'),
+                download_url=parsed_url.get('download'))
 
 
 class PidLockFile(object):