+++ tools/osx/deploy_jenkins_slave.sh
@@ -1,55 +1,71 @@
 #!/bin/sh -eu
-# See tools/posix/deploy_jenkins_slave.sh for more informations and arguments.
+# See tools/posix/deploy_jenkins_slave.sh for more information and arguments.
 
 export OSI="osx"
 
 . "$(python -c "import os.path; print(os.path.realpath('$0').replace('/osx/', '/posix/'))")"
 
-code_sign() {
+prepare_signing() {
     # Code sign the app
-    # TODO: add parameters to Jenkinsfile
     # https://github.com/pyinstaller/pyinstaller/wiki/Recipe-OSX-Code-Signing
-    local app="$1"
 
     if [[ -z ${SIGNING_ID:-} ]]; then
         echo ">>> [sign] WARNING: Signing ID is unavailable, application won't be signed."
         return
-    elif [[ -z ${KEYCHAIN_PWD:-} ]]; then
+    elif [[ -z ${LOGIN_KEYCHAIN_PASSWORD:-} ]]; then
         echo ">>> [sign] WARNING: Keychain is unavailable, application won't be signed."
         return
     fi
 
     echo ">>> [sign] Unlocking the Nuxeo keychain"
-    security unlock-keychain -p "${KEYCHAIN_PWD}" "${KEYCHAIN_PATH}"
-
-    echo ">>> [sign] Signing the file ${app}"
-    codesign --deep --verbose --sign "${SIGNING_ID}" "${app}"
-
-    echo ">>> [sign] Verifying code signature"
-    codesign --verify --verbose "${app}"
+    security unlock-keychain -p "${LOGIN_KEYCHAIN_PASSWORD}" "${LOGIN_KEYCHAIN_PATH}"
+    security set-key-partition-list -S apple-tool:,apple:,codesign: -s -k "${LOGIN_KEYCHAIN_PASSWORD}" "${LOGIN_KEYCHAIN_PATH}"
+    security set-keychain-settings "${LOGIN_KEYCHAIN_PATH}"
 }
 
 create_package() {
     # Create the final DMG
     local app_name="Nuxeo Drive"
-    local bundle_name="Nuxeo Drive.app"
+    local bundle_name="${app_name}.app"
     local output_dir="${WORKSPACE_DRIVE}/dist"
     local pkg_path="${output_dir}/${bundle_name}"
     local src_folder_tmp="${WORKSPACE}/dmg_src_folder.tmp"
     local dmg_tmp="${WORKSPACE}/nuxeo-drive.tmp.dmg"
     local background_file="${WORKSPACE_DRIVE}/tools/osx/dmgbackground.png"
     local plist="${WORKSPACE_DRIVE}/tools/osx/Info.plist"
+    local extension_path="${WORKSPACE_DRIVE}/tools/osx/drive"
     local generated_ds_store="${WORKSPACE_DRIVE}/tools/osx/generated_DS_Store"
     local app_version="$(python "${WORKSPACE_DRIVE}/tools/changelog.py" --drive-version)"
 
     echo ">>> [package] Updating Info.plist"
     sed "s/\$version/${app_version}/" "${plist}" > "${pkg_path}/Contents/Info.plist"
 
+    echo ">>> [package] Building the FinderSync extension"
+    xcodebuild -project "${extension_path}/drive.xcodeproj" -target "NuxeoFinderSync" -configuration Release build
+
+    echo ">>> [package] Adding the extension to the package"
+    mkdir "${pkg_path}/Contents/PlugIns"
+    cp -a  "${extension_path}/build/Release/NuxeoFinderSync.appex" "${pkg_path}/Contents/PlugIns/."
+    rm -rf "${extension_path}/build"
+
     echo ">>> [package] Creating the DMG file"
-    code_sign "${pkg_path}"
+
+    prepare_signing
+    echo ">>> [sign] Signing the app extension"
+    codesign -dfvs "${SIGNING_ID}" --entitlements "${extension_path}/NuxeoFinderSync/NuxeoFinderSync.entitlements" "${pkg_path}/Contents/PlugIns/NuxeoFinderSync.appex"
+
+    echo ">>> [sign] Signing the app"
+    # We recursively sign all the files without --force so the app
+    # extension keeps its entitlements and its sandboxing
+    find "${pkg_path}/Contents/MacOS" -type f -exec codesign -ds "${SIGNING_ID}" {} \;
+    # Then we shallow sign the .app
+    codesign -vs "${SIGNING_ID}" "${pkg_path}"
+
+    echo ">>> [sign] Verifying code signature"
+    codesign -vv "${pkg_path}"
+    spctl --assess -vv "${pkg_path}"
 
     echo ">>> [DMG] ${bundle_name} version ${app_version}"
-
     # Compute DMG name and size
     local dmg_path="${output_dir}/nuxeo-drive-${app_version}.dmg"
     local dmg_size=$(( $(du -sm "${pkg_path}" | cut -d$'\t' -f1,1) + 20 ))
@@ -75,7 +91,8 @@ create_package() {
     # Clean tmp directories
     rm -rf "${src_folder_tmp}" "${dmg_tmp}"
 
-    code_sign "dist/nuxeo-drive-${app_version}.dmg"
+    codesign -vs "${SIGNING_ID}" "dist/nuxeo-drive-${app_version}.dmg"
+    rm -rf "${pkg_path}"
 }
 
 main "$@"