+++ tools/osx/drive/drive/ViewController.swift
@@ -1,0 +1,26 @@
+//
+//  ViewController.swift
+//  drive
+//
+//  Created by Léa Klein on 28/02/2018.
+//  Copyright © 2018 Nuxeo. All rights reserved.
+//
+
+import Cocoa
+
+class ViewController: NSViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override var representedObject: Any? {
+        didSet {
+        // Update the view, if already loaded.
+        }
+    }
+
+}
+